fix: se arregla nombre de carpeta, se mejora onboarding incator
