+++ lib/features/onboarding/screens/components/onboarding_item.dart
