+++ lib/features/onboarding/screens/onboarding_view.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:motow_app/features/onboarding/screens/Components/onboarding_indicator.dart';
 import 'package:motow_app/routing/route_paths.dart';
@@ -8,6 +7,22 @@ import 'Components/onboarding_item.dart';
 
 class OnboardingView extends HookWidget {
   const OnboardingView({Key? key}) : super(key: key);
+
+  final listPage = const [
+    OnboardingItem(
+      secondaryText: 'Encuentra Gruas para moto',
+    ),
+    OnboardingItem(
+      primaryText: 'Guarda, califica y comenta.',
+    ),
+    OnboardingItem(
+      secondaryText: 'Ayuda a la comunidad agregando nuevo datos.',
+    ),
+    OnboardingItem(
+      primaryText: 'Buenas rutas!',
+      blackBackground: true,
+    ),
+  ];
 
   @override
   Widget build(BuildContext context) {
@@ -19,30 +34,17 @@ class OnboardingView extends HookWidget {
         children: [
           PageView(
             controller: pageController,
-            children: const [
-              OnboardingItem(
-                secondaryText: 'Encuentra Gruas para moto',
-              ),
-              OnboardingItem(
-                primaryText: 'Guarda, califica y comenta.',
-              ),
-              OnboardingItem(
-                secondaryText: 'Ayuda a la comunidad agregando nuevo datos.',
-              ),
-              OnboardingItem(
-                primaryText: 'Buenas rutas!',
-                blackBackground: true,
-              ),
-            ],
+            children: listPage,
             onPageChanged: (indexPage) {
               selectedIndex.value = indexPage;
-                if (indexPage == 3) {
-                  Future.delayed(const Duration(milliseconds:1200 )).then(
-                      (value) => Navigator.pushNamed(context, RoutePaths.main));
-                }
+              if (indexPage == 3) {
+                Future.delayed(const Duration(milliseconds: 1200)).then(
+                    (value) => Navigator.pushNamed(context, RoutePaths.main));
+              }
             },
           ),
-          OnboardingIndicator(selectedIndex: selectedIndex.value)
+          OnboardingIndicator(
+              selectedIndex: selectedIndex.value, length: listPage.length)
         ],
       ),
     );
